Fix hdk_setup.sh when downloading a new shell dcp.
- If the download directory doesn't exist then curl fails.
  - mkdir -p if the directory doesn't exist.

+++ hdk_setup.sh
@@ -143,6 +143,9 @@ hdk_shell=$hdk_shell_dir/SH_CL_BB_routed.dcp
 hdk_shell_s3_bucket=aws-fpga-hdk-resources
 s3_hdk_shell=$hdk_shell_s3_bucket/hdk/$hdk_shell_version/build/checkpoints/from_aws/SH_CL_BB_routed.dcp
 # Download the sha256
+if [ ! -e $hdk_shell_dir ]; then
+	mkdir -p $hdk_shell_dir || { err_msg "Failed to create $hdk_shell_dir"; return 2; }
+fi
 # Use curl instead of AWS CLI so that credentials aren't required.
 curl -s https://s3.amazonaws.com/$s3_hdk_shell.sha256 -o $hdk_shell.sha256 || { err_msg "Failed to download HDK shell's checkpoint version from $s3_hdk_shell.sha256 -o $hdk_shell.sha256"; return 2; }
 if grep -q '<?xml version' $hdk_shell.sha256; then
